move secret encryption/storage to Factory
